Allow Job timeout config to be explicitly set to 'never'
This commit adds the configuration option to remove the default timeout behavior
of setting a Job's timeout to the interval, if interval is set. A job can now
have their timeout explicitly removed without setting a default. This helps
scenarios where you want periodic jobs fired on an interval but not SIGKILL'ing
the previous job and instead waiting for them to complete.

NOTE: This patch is missing the ability for the periodic job timer to pause
while a job is still running.

Ref: #525